rego system and credit system integration

+++ Models/MasterDbModels/Companies.cs
@@ -7,6 +7,6 @@ public class Companies
     [Key]
     public int _id { get; set; }
     public string CompanyID { get; set; }
-    public int Credits { get; set; }
+    public double Credits { get; set; }
     
 }
